TIKA-2368 -- move to different package to avoid split package warning
