+++ tika-parsers/src/test/java/org/apache/tika/parser/sentiment/SentimentAnalysisParserTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.tika.parser.sentiment.analysis;
+package org.apache.tika.parser.sentiment;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
