+++ tika-parsers/src/test/resources/org/apache/tika/parser/sentiment/tika-config-sentiment-opennlp-cat.xml
@@ -16,7 +16,7 @@
 -->
 <properties>
     <parsers>
-        <parser class="org.apache.tika.parser.sentiment.analysis.SentimentAnalysisParser">
+        <parser class="org.apache.tika.parser.sentiment.SentimentAnalysisParser">
             <mime>text/plain</mime>
             <mime>application/sentiment</mime>
             <params>
